+++ integration/frameworks/vue/vue-procaptcha-integration-demo/vite.cjs.config.ts
@@ -1,3 +1,4 @@
+import path from "node:path";
 // Copyright 2021-2025 Prosopo (UK) Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,13 +12,15 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import { assert, describe, test } from "vitest";
+import { ViteCommonJSConfig } from "@prosopo/config";
+import vue from "@vitejs/plugin-vue";
 
-describe("dummy", () => {
-	test("dummy", async () => {
-		assert(
-			true,
-			"This is a dummy test to ensure vite doesn't complain about no tests found.",
-		);
-	}, 120000);
-});
+export default async function () {
+	const config = await ViteCommonJSConfig(
+		path.basename("."),
+		path.resolve("./tsconfig.json"),
+	);
+	config.plugins = config.plugins || [];
+	config.plugins.push(vue());
+	return config;
+}