+++ packages/user-access-policy/src/api/insertRulesEndpoint.ts
@@ -30,12 +30,12 @@ export const insertRulesEndpointSchema: z.ZodType<{
 	accessPolicy: z.infer<typeof accessPolicySchema>;
 	policyScope?: z.infer<typeof policyScopeSchema>;
 	userScopes: z.input<typeof userScopeInputSchema>[];
-	expirationTimestampSeconds?: number;
+	expirationTimestamp?: number;
 }> = z.object({
 	accessPolicy: accessPolicySchema,
 	policyScope: policyScopeSchema.optional(),
 	userScopes: z.array(userScopeInputSchema),
-	expirationTimestampSeconds: z
+	expirationTimestamp: z
 		.number()
 		.optional()
 		.transform((val) => (val !== undefined ? Math.floor(val) : val)),
@@ -109,10 +109,7 @@ export class InsertRulesEndpoint
 			};
 
 			createPromises.push(
-				this.accessRulesWriter.insertRule(
-					rule,
-					args.expirationTimestampSeconds,
-				),
+				this.accessRulesWriter.insertRule(rule, args.expirationTimestamp),
 			);
 		}
 		return Promise.all(createPromises);