+++ demos/provider-mock/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prosopo/provider-mock
 
+## 2.7.18
+### Patch Changes
+
+- Updated dependencies [b7c3258]
+  - @prosopo/provider@3.2.0
+
+## 2.7.17
+### Patch Changes
+
+  - @prosopo/provider@3.1.3
+
+## 2.7.16
+### Patch Changes
+
+  - @prosopo/provider@3.1.2
+
 ## 2.7.15
 ### Patch Changes