+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @prosopo/provider
 
+## 3.2.0
+### Minor Changes
+
+- b7c3258: Add tests for UAPs
+
+### Patch Changes
+
+- Updated dependencies [b7c3258]
+  - @prosopo/user-access-policy@3.3.0
+  - @prosopo/env@3.1.0
+  - @prosopo/database@3.0.9
+  - @prosopo/types-database@3.0.9
+  - @prosopo/datasets@3.0.9
+  - @prosopo/types-env@2.7.13
+
+## 3.1.3
+### Patch Changes
+
+- Updated dependencies [cdf7c29]
+  - @prosopo/user-access-policy@3.2.1
+  - @prosopo/database@3.0.8
+  - @prosopo/types-database@3.0.8
+  - @prosopo/env@3.0.8
+  - @prosopo/datasets@3.0.8
+  - @prosopo/types-env@2.7.12
+
+## 3.1.2
+### Patch Changes
+
+- Updated dependencies [a7164ce]
+  - @prosopo/user-access-policy@3.2.0
+  - @prosopo/database@3.0.7
+  - @prosopo/types-database@3.0.7
+  - @prosopo/env@3.0.7
+  - @prosopo/datasets@3.0.7
+  - @prosopo/types-env@2.7.11
+
 ## 3.1.1
 ### Patch Changes